+++ packages/orq-rc/RELEASES.md
@@ -12619,3 +12619,13 @@ Based on:
 - [typescript v4.16.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.16.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.17 - packages/orq-rc
+
+## 2026-09-25 07:44:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.799.0 (2.941.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.16.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v4.16.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.18 - packages/orq-rc